acloud: Add in AVDSpec class and basic arg parsing.
Add in a very barebones implementation of AVDSpec. The real meat of the
class will come later once we add in the new options (e.g. --local_instance)
and we do the magic for our users.

I intentionally did not do any of the remote image arg inferring logic
as that will require a new arg and some google api calls (to Android
Build) so that deserves its own cl.

Bug: 112598956
Test: m acloud && acloud create
m acloud_test && acloud_test

$ acloud create --build_id 1234 --branch rand_branch
We will (but not yet) create an AVD with these details:
 - avd type: cuttlefish
  - num of instances requested: 1
   - remote image details: {'build_id': '1234', 'build_branch':
   'git_rand_branch'}.

Change-Id: Ia27bf86e55f3a171ba39663bb2c1db407119a54f