acloud create local instance tweaks.
1. If the user just setup their workstation and started up a local
instance and then wanted to start another local instance before stopping
the other one, acloud will try to stop the current running instance.
However since the user just setup their host, we need to pass in all the
user groups to the command so it succeeds, we weren't so the attempt to
stop the instance failed all the time, now we succeed!

2. The pre-run check use to check if the user's host had packages
installed AND if they were part of the user groups, and if either was
not true, we'd have them go through the setup process. This meant that a
user who just ran setup on their workstation will have to run through
setup _every_ time they run a local instance until they logged out. That
sucks so updated the pre-run check to only check if packages are
installed and take that as a signal that the user has been added to the
groups and we'll take care of it at execution time.

Bug: 117615246
Test: acloud create --local_instance && acloud create --local_instance
and see acloud properly restart the instance.

Change-Id: If872965329224748b6a3f3181be88d378575ff69