Add support for tags
Goldfish is getting ready to release base-images that expose https
endpoints. This means that we need to configure the firewall to allow
https to go through.

Opening up the firewall is done by setting tags on the instance, these
tags can be provided from the command line. For example:

$ acloud create_gf --tags http-server https-server ....

Will create an instance with the firewall opened for http and https.

Includes fixes to properly run unit-test that can be run from setup.py.

Change-Id: I8344c21c22f0d2760fc73529df0810a70797c08a