cuttlefish: Only upload the images that are required.
The super.img image is formed of multiple partitions whose contents
are the same as some of the other images, such as system.img and
vendor.img. Cuttlefish only needs super.img and not any of the
partition images. Therefore, simply uploading *.img results in
uploading the same data twice needlessly.

This change makes it so that we only upload the images that
are required in order to boot. This results in a reduction
of time spent processing and uploading local images (for
aosp_cf_x86_64_phone-userdebug) from 54s to 32s on my machine, and
the size of the compressed uploaded tarball goes from 2.2GB to 1.1GB.

Change-Id: I1e28ddc890e1993c8e3125de206a805fc4cd2885